move 'differenceWith' tests to separate file
